fix(libcore): drop unused fmt import (build break) + reviewer follow-ups
The unused "fmt" import in libcore/io.go failed gomobile compilation,
which broke all four Depot CI jobs (build/test/lint/instrumented depend
on libcore-from-source).

Also addresses CodeRabbit/Greptile findings on PR #108:
- http.go: bound the full request (incl. body reads) with a client
  timeout on the h1h2 client and both doH3Direct clients (ECH + H3).
- http.go/io.go: drop world-writable Chmod(0666) on temp files;
  os.CreateTemp already creates them 0600 and they are atomically
  renamed (over-permissive on per-app-UID Android).
- http.go: remove dead contentLimit field.
- SubscriptionUpdater: use Long arithmetic in the due-check to avoid
  Int overflow, matching computeSubscriptionWorkSchedule.
- BackupFragment: require WebDAV PROPFIND 207 before parsing the
  multistatus body (a 200/204 proxy/login page no longer falls through).
- AssetsActivity: drop dead post-unxz size check (Libcore.unxz already
  enforces the same 256 MB cap and fails before rename).
- .depot/workflows: make libcore/sidecars cache fingerprints content-
  and path-aware via 'git ls-files -s | sha256sum' (symlink-safe; the
  old find|cat masked content moves and a dangling buildScript/nkmr
  symlink), normalize the libcore-status command identically across all
  four workflows (shared cache key), and pass KEYSTORE_B64 via env:.

+++ app/src/main/java/io/nekohasekai/sagernet/bg/SubscriptionUpdater.kt
@@ -110,10 +110,11 @@ object SubscriptionUpdater {
             var attempted = false
             var failed = false
             if (subscriptions.isNotEmpty()) {
-                val nowSeconds = (System.currentTimeMillis() / 1000).toInt()
+                val nowSeconds = System.currentTimeMillis() / 1000L
                 for ((profile, subscription) in subscriptions) {
-                    val lastUpdated = subscription.lastUpdated ?: 0
-                    val delaySeconds = (subscription.autoUpdateDelay ?: 1440).coerceAtLeast(15) * 60
+                    val lastUpdated = (subscription.lastUpdated ?: 0).toLong()
+                    val delaySeconds =
+                        (subscription.autoUpdateDelay ?: 1440).toLong().coerceAtLeast(15L) * 60L
                     if (nowSeconds - lastUpdated < delaySeconds) {
                         Logs.d("work: not updating " + profile.displayName())
                         continue
